# docs/08-openapi.yaml
openapi: 3.1.0
info:
  title: Velora Platform API
  version: 0.1.0
  description: Production contract baseline. Current repository implements health and demo compliance status only.
servers:
  - url: /api
paths:
...
  /compliance/status:
    get:
      operationId: getComplianceStatus
      parameters:
        - in: query
          name: userId
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Purpose-limited capability status }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-velora_session
...
  responses:
    BadRequest: { description: Invalid request }
...